Create views for habit list and habit card
